fix adding words

+++ src/main/java/by/bntu/fitr/poisit/lytkina/englishLearnBot/repo/WordRepoI.java
@@ -3,5 +3,5 @@ package by.bntu.fitr.poisit.lytkina.englishLearnBot.repo;
 import by.bntu.fitr.poisit.lytkina.englishLearnBot.bean.Word;
 import org.springframework.data.repository.CrudRepository;
 
-public interface WordRepoI extends CrudRepository<Word, Integer> {
+public interface WordRepoI extends CrudRepository<Word, Long> {
 }
